+++ Sources/KeyboardKit/Feedback/StandardKeyboardFeedbackHandler.swift
@@ -29,12 +29,10 @@ open class StandardKeyboardFeedbackHandler: KeyboardFeedbackHandler {
         self.settings = settings
     }
     
-    
     public let settings: KeyboardFeedbackSettings
     
     public var audioConfig: AudioFeedbackConfiguration { settings.audioConfiguration }
     public var hapticConfig: HapticFeedbackConfiguration { settings.hapticConfiguration }
-    
     
     /**
      Whether or not a feedback should be given for a certain
@@ -90,6 +88,8 @@ open class StandardKeyboardFeedbackHandler: KeyboardFeedbackHandler {
      certain `action`.
      */
     open func triggerAudioFeedback(for gesture: KeyboardGesture, on action: KeyboardAction) {
+        let custom = audioConfig.actions.first { $0.action == action }
+        if let custom = custom { return custom.feedback.play() }
         if action == .backspace { return audioConfig.delete.play() }
         if action.isInputAction { return audioConfig.input.play() }
         if action.isSystemAction { return audioConfig.system.play() }
@@ -100,6 +100,8 @@ open class StandardKeyboardFeedbackHandler: KeyboardFeedbackHandler {
      certain `action`.
      */
     open func triggerHapticFeedback(for gesture: KeyboardGesture, on action: KeyboardAction) {
+        let custom = hapticConfig.actions.first { $0.action == action && $0.gesture == gesture }
+        if let custom = custom { return custom.feedback.trigger() }
         switch gesture {
         case .doubleTap: hapticConfig.doubleTap.trigger()
         case .longPress: hapticConfig.longPress.trigger()